refactor: move tools/ghx packages to pkg/ghx
No need to keep remaining packages separate. They can be
 merged under ghx
