+++ cmd/ghx/README.md
@@ -3,8 +3,6 @@
 The GitHub Actions Executor is a helper command for running [GitHub Actions](https://docs.github.com/en/actions) from
 inside of [Dagger](https://dagger.io) containers.
 
-
-
 ## Usage
 
 ### Global Configuration
