+++ pkg/ghx/container_executor.go
@@ -1,4 +1,4 @@
-package runner
+package ghx
 
 import (
 	"bufio"
@@ -14,7 +14,6 @@ import (
 	"github.com/aweris/gale/internal/core"
 	"github.com/aweris/gale/internal/expression"
 	"github.com/aweris/gale/internal/log"
-	"github.com/aweris/gale/tools/ghx/actions"
 )
 
 var _ Executor = new(ContainerExecutor)
@@ -25,8 +24,8 @@ type ContainerExecutor struct {
 	entrypoint string                  // entrypoint is the entrypoint of the container
 	args       []string                // args is the arguments of the container
 	env        map[string]string       // env to pass to the command as environment variables
-	ec         *actions.ExprContext    // ec is the expression context to evaluate the github expressions
-	dec        *actions.ExprContext    // dec is the dagger expression context to evaluate action meta files. TODO: temporary solution, we need to find a better way to do this
+	ec         *ExprContext            // ec is the expression context to evaluate the github expressions
+	dec        *ExprContext            // dec is the dagger expression context to evaluate action meta files. TODO: temporary solution, we need to find a better way to do this
 	commands   []*core.WorkflowCommand // commands is the list of commands that are executed in the step
 	envFiles   *EnvironmentFiles       // envFiles contains temporary files that can be used to perform certain actions
 }
