+++ pkg/ghx/context.go
@@ -1,4 +1,4 @@
-package actions
+package ghx
 
 import (
 	"fmt"
