+++ pkg/ghx/runner_step.go
@@ -1,4 +1,4 @@
-package runner
+package ghx
 
 import (
 	"context"
